[UE model] 64bit support in ASTF programming model

+++ scripts/automation/trex_control_plane/interactive/trex/astf/trex_astf_profile.py
@@ -626,7 +626,7 @@ class ASTFProgram(object):
                   var_id  : string
                      var-id there are limited number of variables 
 
-                  val  : uint32_t
+                  val  : uint64_t
                      value of the variable 
 
         """
